+++ RabbitMqHttpApiClient/Models/Common/MessageStatsModel/IGetNoAck.cs
@@ -5,7 +5,7 @@ namespace RabbitMqHttpApiClient.Models.Common.MessageStatsModel
     public interface IGetNoAck
     {
         [JsonProperty("get_no_ack")]
-        int GetNoAck { get; set; }
+        long GetNoAck { get; set; }
 
         [JsonProperty("get_no_ack_details")]
         RateDetails GetNoAckDetails { get; set; }
