+++ RabbitMqHttpApiClient/Models/Common/MessageStatsModel/IRedeliver.cs
@@ -5,7 +5,7 @@ namespace RabbitMqHttpApiClient.Models.Common.MessageStatsModel
     public interface IRedeliver
     {
         [JsonProperty("redeliver")]
-        int Redeliver { get; set; }
+        long Redeliver { get; set; }
 
         [JsonProperty("redeliver_details")]
         RateDetails RedeliverDetails { get; set; }
