+++ RabbitMqHttpApiClient/Models/QueueModel/GetQueueMessagesRequest.cs
@@ -2,7 +2,7 @@ namespace RabbitMqHttpApiClient.Models.QueueModel
 {
     public class GetQueueMessagesRequest 
     {
-        public int count { get; set; }
+        public long count { get; set; }
         public bool requeue { get; set; }
         public string encoding { get; set; }
     }
